Create prediction_script.py
The input file is a file containing smiles and can be a single column file with a column named 'SMILES'. The default file name is 'input_file.csv'. The QED values will be calculated and deposited in the output file which could help in selecting the best subset of ligands to move forward. Please note that the model was built around the phenoxy-methyl biphenyl core scaffold and should be ideally be used for the same class of molecules.